+++ bitalarm-impl/src/main/java/com/dwarfeng/bitalarm/impl/cache/EnabledAlarmSettingCacheImpl.java
@@ -5,6 +5,7 @@ import com.dwarfeng.bitalarm.stack.bean.entity.AlarmSetting;
 import com.dwarfeng.bitalarm.stack.cache.EnabledAlarmSettingCache;
 import com.dwarfeng.subgrade.impl.cache.RedisKeyListCache;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
 import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
 import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
 import com.dwarfeng.subgrade.stack.exception.CacheException;
@@ -43,6 +44,7 @@ public class EnabledAlarmSettingCacheImpl implements EnabledAlarmSettingCache {
 
     @Override
     @BehaviorAnalyse
+    @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public List<AlarmSetting> get(LongIdKey key) throws CacheException {
         return delegate.get(key);
@@ -50,6 +52,7 @@ public class EnabledAlarmSettingCacheImpl implements EnabledAlarmSettingCache {
 
     @Override
     @BehaviorAnalyse
+    @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public List<AlarmSetting> get(LongIdKey key, int beginIndex, int maxEntity) throws CacheException {
         return delegate.get(key, beginIndex, maxEntity);
@@ -57,6 +60,7 @@ public class EnabledAlarmSettingCacheImpl implements EnabledAlarmSettingCache {
 
     @Override
     @BehaviorAnalyse
+    @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public List<AlarmSetting> get(LongIdKey key, PagingInfo pagingInfo) throws CacheException {
         return delegate.get(key, pagingInfo);
@@ -65,21 +69,21 @@ public class EnabledAlarmSettingCacheImpl implements EnabledAlarmSettingCache {
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
-    public void set(LongIdKey key, Collection<AlarmSetting> entities, long timeout) throws CacheException {
+    public void set(LongIdKey key, @SkipRecord Collection<AlarmSetting> entities, long timeout) throws CacheException {
         delegate.set(key, entities, timeout);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
-    public void leftPush(LongIdKey key, Collection<AlarmSetting> entities, long timeout) throws CacheException {
+    public void leftPush(LongIdKey key, @SkipRecord Collection<AlarmSetting> entities, long timeout) throws CacheException {
         delegate.leftPush(key, entities, timeout);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
-    public void rightPush(LongIdKey key, Collection<AlarmSetting> entities, long timeout) throws CacheException {
+    public void rightPush(LongIdKey key, @SkipRecord Collection<AlarmSetting> entities, long timeout) throws CacheException {
         delegate.rightPush(key, entities, timeout);
     }