+++ bitalarm-impl/src/main/java/com/dwarfeng/bitalarm/impl/dao/AlarmTypeIndicatorDaoImpl.java
@@ -7,6 +7,7 @@ import com.dwarfeng.subgrade.impl.dao.HibernateBatchBaseDao;
 import com.dwarfeng.subgrade.impl.dao.HibernateEntireLookupDao;
 import com.dwarfeng.subgrade.sdk.bean.key.HibernateStringIdKey;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
 import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.DaoException;
@@ -68,48 +69,51 @@ public class AlarmTypeIndicatorDaoImpl implements AlarmTypeIndicatorDao {
 
     @Override
     @BehaviorAnalyse
+    @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
-    public List<StringIdKey> batchInsert(List<AlarmTypeIndicator> elements) throws DaoException {
+    public List<StringIdKey> batchInsert(@SkipRecord List<AlarmTypeIndicator> elements) throws DaoException {
         return batchBaseDao.batchInsert(elements);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
-    public void batchUpdate(List<AlarmTypeIndicator> elements) throws DaoException {
+    public void batchUpdate(@SkipRecord List<AlarmTypeIndicator> elements) throws DaoException {
         batchBaseDao.batchUpdate(elements);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
-    public void batchDelete(List<StringIdKey> keys) throws DaoException {
+    public void batchDelete(@SkipRecord List<StringIdKey> keys) throws DaoException {
         batchBaseDao.batchDelete(keys);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
-    public boolean allExists(List<StringIdKey> keys) throws DaoException {
+    public boolean allExists(@SkipRecord List<StringIdKey> keys) throws DaoException {
         return batchBaseDao.allExists(keys);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
-    public boolean nonExists(List<StringIdKey> keys) throws DaoException {
+    public boolean nonExists(@SkipRecord List<StringIdKey> keys) throws DaoException {
         return batchBaseDao.nonExists(keys);
     }
 
     @Override
     @BehaviorAnalyse
+    @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
-    public List<AlarmTypeIndicator> batchGet(List<StringIdKey> keys) throws DaoException {
+    public List<AlarmTypeIndicator> batchGet(@SkipRecord List<StringIdKey> keys) throws DaoException {
         return batchBaseDao.batchGet(keys);
     }
 
     @Override
     @BehaviorAnalyse
+    @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public List<AlarmTypeIndicator> lookup() throws DaoException {
         return entireLookupDao.lookup();
@@ -117,6 +121,7 @@ public class AlarmTypeIndicatorDaoImpl implements AlarmTypeIndicatorDao {
 
     @Override
     @BehaviorAnalyse
+    @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public List<AlarmTypeIndicator> lookup(PagingInfo pagingInfo) throws DaoException {
         if (pagingInfo.getRows() < 0) {